JWT RS256 Auth

+++ src/ITLabReports/api/server/handlers.go
@@ -1,8 +1,8 @@
 package server
 
 import (
-	"../models"
-	"../utils"
+	"ITLabReports/models"
+	"ITLabReports/utils"
 	"context"
 	"encoding/json"
 	"github.com/gorilla/mux"
@@ -16,9 +16,32 @@ import (
 
 func getAllReports(w http.ResponseWriter, r *http.Request) {
 	reports := make([]models.Report, 0)
+	var filter bson.M
+
 	w.Header().Set("Content-Type", "application/json")
+	roleClaim, err := getClaim(r, "role")
+	if err != nil {
+		utils.AuthError(w, err)
+		return
+	}
+	subClaim, err := getClaim(r, "sub")
+	if err != nil {
+		utils.AuthError(w, err)
+		return
+	}
+
+	switch roleClaim {
+	case "admin":
+		filter = bson.M{}
+	case "user":
+		filter = bson.M{"reportsender": subClaim}
+	default:
+		w.WriteHeader(403)
+		w.Write([]byte("wrong role claim"))
+		return
+	}
 	ctx, _ := context.WithTimeout(context.Background(), 10*time.Second)
-	cur, err := collection.Find(ctx, bson.D{})
+	cur, err := collection.Find(ctx, filter)
 	if err != nil {
 		log.Panic(err)
 	}
@@ -32,8 +55,30 @@ func getAllReports(w http.ResponseWriter, r *http.Request) {
 	json.NewEncoder(w).Encode(reports)
 }
 func getAllReportsSorted(w http.ResponseWriter, r *http.Request) {
+	var filter bson.M
 	reports := make([]models.Report, 0)
 	w.Header().Set("Content-Type", "application/json")
+	roleClaim, err := getClaim(r, "role")
+	if err != nil {
+		utils.AuthError(w, err)
+		return
+	}
+	subClaim, err := getClaim(r, "sub")
+	if err != nil {
+		utils.AuthError(w, err)
+		return
+	}
+	switch roleClaim {
+	case "admin":
+		filter = bson.M{}
+	case "user":
+		filter = bson.M{"reportsender": subClaim}
+	default:
+		w.WriteHeader(403)
+		w.Write([]byte("wrong role claim"))
+		return
+	}
+
 	ctx, _ := context.WithTimeout(context.Background(), 10*time.Second)
 	data := mux.Vars(r)
 	sortVar := data["var"]
@@ -45,7 +90,7 @@ func getAllReportsSorted(w http.ResponseWriter, r *http.Request) {
 		findOptions.SetSort(bson.M{"date": 1})
 	}
 
-	cur, err := collection.Find(ctx, bson.D{}, findOptions)
+	cur, err := collection.Find(ctx, filter, findOptions)
 	if err != nil {
 		log.Panic(err)
 	}
@@ -59,8 +104,10 @@ func getAllReportsSorted(w http.ResponseWriter, r *http.Request) {
 	json.NewEncoder(w).Encode(reports)
 }
 func getReport(w http.ResponseWriter, r *http.Request) {
+	var filter bson.M
 	var report models.Report
 	w.Header().Set("Content-Type", "application/json")
+
 	json.NewDecoder(r.Body).Decode(&report)
 	data := mux.Vars(r)
 
@@ -69,19 +116,35 @@ func getReport(w http.ResponseWriter, r *http.Request) {
 		http.NotFound(w, r)
 		return
 	}
-	filter := bson.M{"_id": objID}
-	ctx, _ := context.WithTimeout(context.Background(), 10*time.Second)
 
-	ctx, _ = context.WithTimeout(context.Background(), 10*time.Second)
+	roleClaim, err := getClaim(r, "role")
+	if err != nil {
+		utils.AuthError(w, err)
+		return
+	}
+	subClaim, err := getClaim(r, "sub")
+	if err != nil {
+		utils.AuthError(w, err)
+		return
+	}
+	filter = bson.M{"_id": objID}
+
+
+	ctx, _ := context.WithTimeout(context.Background(), 10*time.Second)
 	err = collection.FindOne(ctx, filter).Decode(&report)
 	if err != nil {
 		http.NotFound(w, r)
 		return
 	}
-	json.NewEncoder(w).Encode(report)
+	if report.ReportSender == subClaim || roleClaim == "admin" {
+		json.NewEncoder(w).Encode(report)
+	} else {
+		w.WriteHeader(403)
+		return
+	}
 }
 func getEmployeeSample(w http.ResponseWriter, r *http.Request) {
-	//query string: db.reports.find({"reportsender":"Anton", "$and" : [{"date" : {"$gte":"2019-12-31"}}, {"date" : {"$lte" : "2020-01-06"}} ]})
+	var filter bson.D
 	reports := make([]models.Report, 0)
 	w.Header().Set("Content-Type", "application/json")
 
@@ -89,16 +152,31 @@ func getEmployeeSample(w http.ResponseWriter, r *http.Request) {
 	employee := data["employee"]
 	dateBegin := utils.FormatQueryDate(data["dateBegin"])+"T00:00:00"
 	dateEnd := utils.FormatQueryDate(data["dateEnd"])+"T23:59:59"
-	log.Println(dateBegin)
-	log.Println(dateEnd)
 	findOptions := options.Find().SetSort(bson.M{"date": 1})
-	filter := bson.D{
-		{"reportsender" ,employee},
-		{"$and", []interface{}{
-			bson.D{{"date",bson.M{"$gte": dateBegin}}},
-			bson.D{{"date", bson.M{"$lte" : dateEnd}}},
-		}},
+
+	roleClaim, err := getClaim(r, "role")
+	if err != nil {
+		utils.AuthError(w, err)
+		return
 	}
+	subClaim, err := getClaim(r, "sub")
+	if err != nil {
+		utils.AuthError(w, err)
+		return
+	}
+	if employee == subClaim || roleClaim == "admin" {
+		filter = bson.D{
+			{"reportsender" ,employee},
+			{"$and", []interface{}{
+				bson.D{{"date",bson.M{"$gte": dateBegin}}},
+				bson.D{{"date", bson.M{"$lte" : dateEnd}}},
+			}},
+		}
+	} else {
+		w.WriteHeader(403)
+		return
+	}
+
 	ctx, _ := context.WithTimeout(context.Background(), 10*time.Second)
 	cur, err := collection.Find(ctx, filter, findOptions)
 	if err != nil {
@@ -112,14 +190,20 @@ func getEmployeeSample(w http.ResponseWriter, r *http.Request) {
 	if err != nil {
 		log.Fatal(err)
 	}
-	log.Println(len(reports))
 	json.NewEncoder(w).Encode(reports)
 }
 func createReport(w http.ResponseWriter, r *http.Request) {
 	var report models.Report
 	w.Header().Set("Content-Type", "application/json")
 	json.NewDecoder(r.Body).Decode(&report)
-	report.ReportSender = r.Header.Get("CustomAuthor")
+
+	subClaim, err := getClaim(r, "sub")
+	if err != nil {
+		utils.AuthError(w, err)
+		return
+	}
+	report.ReportSender = subClaim
+
 	headerDate := r.Header.Get("Date")
 	report.Date = utils.FormatDate(headerDate)
 
@@ -147,24 +231,42 @@ func updateReport(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 	filter := bson.M{"_id": objID}
-	ctx, _ := context.WithTimeout(context.Background(), 10*time.Second)
 
-	ctx, _ = context.WithTimeout(context.Background(), 10*time.Second)
+	ctx, _ := context.WithTimeout(context.Background(), 10*time.Second)
 	err = collection.FindOne(ctx, filter).Decode(&updatedReport)
 	if err != nil {
 		http.NotFound(w, r)
 		return
 	}
-	updatedReport.Text = report.Text
-	updateResult, err := collection.ReplaceOne(ctx, filter, updatedReport)
-	if err != nil || updateResult.MatchedCount == 0 {
-		http.NotFound(w, r)
+	roleClaim, err := getClaim(r, "role")
+	if err != nil {
+		utils.AuthError(w, err)
 		return
 	}
-	json.NewEncoder(w).Encode(updatedReport)
+	subClaim, err := getClaim(r, "sub")
+	if err != nil {
+		utils.AuthError(w, err)
+		return
+	}
+
+	if updatedReport.ReportSender == subClaim || roleClaim == "admin" {
+		updatedReport.Text = report.Text
+		updateResult, err := collection.ReplaceOne(ctx, filter, updatedReport)
+		if err != nil || updateResult.MatchedCount == 0 {
+			http.NotFound(w, r)
+			return
+		}
+		json.NewEncoder(w).Encode(updatedReport)
+	} else {
+		w.WriteHeader(403)
+		return
+	}
+
 }
 func deleteReport(w http.ResponseWriter, r *http.Request) {
+	var report models.Report
 	w.Header().Set("Content-Type", "application/json")
+
 	data := mux.Vars(r)
 	objID, err := primitive.ObjectIDFromHex(string(data["id"]))
 	if err != nil {
@@ -173,10 +275,32 @@ func deleteReport(w http.ResponseWriter, r *http.Request) {
 	}
 	filter := bson.M{"_id": objID}
 	ctx, _ := context.WithTimeout(context.Background(), 10*time.Second)
-	deleteResult, err := collection.DeleteOne(ctx, filter)
-	if err != nil || deleteResult.DeletedCount == 0 {
+	err = collection.FindOne(ctx, filter).Decode(&report)
+	if err != nil {
 		http.NotFound(w, r)
 		return
 	}
-	w.WriteHeader(200)
+	roleClaim, err := getClaim(r, "role")
+	if err != nil {
+		utils.AuthError(w, err)
+		return
+	}
+	subClaim, err := getClaim(r, "sub")
+	if err != nil {
+		utils.AuthError(w, err)
+		return
+	}
+	if report.ReportSender == subClaim || roleClaim == "admin" {
+		ctx, _ = context.WithTimeout(context.Background(), 10*time.Second)
+		deleteResult, err := collection.DeleteOne(ctx, filter)
+		if err != nil || deleteResult.DeletedCount == 0 {
+			http.NotFound(w, r)
+			return
+		}
+		w.WriteHeader(200)
+	} else {
+		w.WriteHeader(403)
+		return
+	}
 }
+
